2, 3. KochFractalAction: remove INeed using, pass IImageHolder and Pallette via DI-container
