+++ FractalPainter/App/Actions/KochFractalAction.cs
@@ -6,19 +6,15 @@ using Ninject;
 
 namespace FractalPainting.App.Actions
 {
-    public class KochFractalAction : IUiAction, INeed<IImageHolder>, INeed<Palette>
+    public class KochFractalAction : IUiAction
     {
         private IImageHolder imageHolder;
         private Palette palette;
 
-        public void SetDependency(IImageHolder dependency)
+        public KochFractalAction(IImageHolder imageHolder, Palette palette)
         {
-            imageHolder = dependency;
-        }
-
-        public void SetDependency(Palette dependency)
-        {
-            palette = dependency;
+            this.imageHolder = imageHolder;
+            this.palette = palette;
         }
 
         public string Category => "Фракталы";
