+++ FractalPainter/App/MainForm.cs
@@ -11,32 +11,21 @@ namespace FractalPainting.App
 {
     public class MainForm : Form
     {
-        public MainForm(IUiAction[] actions)
+        public MainForm(IUiAction[] actions, IImageHolder imageHolder, Palette palette, SettingsManager settingsManager)
         {
-            var imageSettings = CreateSettingsManager().Load().ImageSettings;
+            var imageSettings = settingsManager.Load().ImageSettings;
             ClientSize = new Size(imageSettings.Width, imageSettings.Height);
 
             var mainMenu = new MenuStrip();
             mainMenu.Items.AddRange(actions.ToMenuItems());
             Controls.Add(mainMenu);
 
-            var pictureBox = new PictureBoxImageHolder();
-            pictureBox.RecreateImage(imageSettings);
-            pictureBox.Dock = DockStyle.Fill;
-            Controls.Add(pictureBox);
-
-            DependencyInjector.Inject<IImageHolder>(actions, pictureBox);
-            DependencyInjector.Inject<IImageDirectoryProvider>(actions, CreateSettingsManager().Load());
-            DependencyInjector.Inject<IImageSettingsProvider>(actions, CreateSettingsManager().Load());
-            DependencyInjector.Inject(actions, new Palette());
-        }
-
-        private static SettingsManager CreateSettingsManager()
-        {
-            var container = new StandardKernel();
-            container.Bind<IObjectSerializer>().To<XmlObjectSerializer>();
-            container.Bind<IBlobStorage>().To<FileBlobStorage>();
-            return container.Get<SettingsManager>();
+            Controls.Add(imageHolder as PictureBox);
+ 
+            DependencyInjector.Inject<IImageHolder>(actions, imageHolder);
+            DependencyInjector.Inject<IImageDirectoryProvider>(actions, settingsManager.Load());
+            DependencyInjector.Inject<IImageSettingsProvider>(actions, settingsManager.Load());
+            DependencyInjector.Inject(actions, palette);
         }
 
         protected override void OnShown(EventArgs e)
